Correct the real firstTid in _bitmap_union
In previous refactor which try to resolve issue:
We change the way to calculate the firstTid holding in the
current position batch words, since concurrent insert will cause
the wrong tid matching in words_get_match().

To fix that issue in PR #11377,
we always get the correct firstTid from the bitmap index page.
But _bitmap_union is not touched. It was responsible to union bitmap
batch words for multivalues index scan like `where x in (x1, x2)` or `x > v`.

And also the logic for the catch up loop is not correct in
_bitmap_catchup_to_next_tid, it should check remain words number.