Fix flaky failed cases pg_rewind_fail_missing_xlog.
The case is tesing the bug when gp crash recovery, it calls
two checkpoints, if some checkpoint triggered removed/recylced
wal file, the gp_rewind might failed.

The flaky is due to sometimes when stoping the primary and
promoting the mirror, the end of recovery hasn't fully completed
after mirror is promoted, then insert some data will see errors
"seg is in recovery mode".

The fix is waiting for the end of recovery CHECKPOINT completed
after the mirror was promoted.