Support backward scanning of spilled files.
In the previous commit a436c5a11a91f27e13f09260439f3bc8ad7f777f
After the tuple there is also a tuple length,called trailing length,
for backward scanning.  The problem is that the trailing length
was checked for the leading bit, which does not contain this bit at all.

This commit is adding leading bit in the trailing length.