Fix the potential risk for using erreport.
After adding fault inject tests, we found that sometimes it will panic.
The stack trace shows that rxThreadFunc()->handleMismatch()->sendAck()
->sendControlMessage()->sendtoWithRetry(), which panics at EmitErrorReport.
Because ereport is NOT thread-safe.

This part of logic is introduced in the PR https://github.com/greenplum-db/gpdb/pull/13371.
Before that ereport is safe to use in sendtoWithRetry() because it won't
be used in the threadFunc.

So revert it to sendto() in sendControlMessage() for now.