Improve upgrade workflow.
1. running the command "el8_migrate_locale" after sourcing "greenplum_path.sh",
no need to specify "python" nor a ".py" file extension to execute the utility.

2. execute "el8_migrate_locale identify" run a check of an existing Greenplum
cluster for impacted index and tables, and it will output to STDOUT information.

3. execute "el8_migrate_locale prepare --output [FILE]" prepare for an upgrade,
and it will output to the given [FILE] the commands for migrating BOTH the
impacted index and partitioned tables.

4. execute "el8_migrate_locale validate" will validate after migration that
data correctness of partitioned tables.